Specify validations on User model
The User model should validate for the presence of its fields.

This test doesn't get slowed down by testing for validations on every single field; it tests a few fields and reasonably assumes that if some validations are working and nobody is deleting validation code in the model, the other validations would be working as well.