Implement a User model
This commit implements a User model with a role attribute that can currently only have the value "admin", as well as name attributes, an email attribute, and password and password confirmation attributes at instantiation that correspond to a password_digest column in the database. The User model validates for the presence of all of its attributes. A user instance can be created like:

User.create(role: "admin",
            first_name: "Firstname",
            last_name: "Lastname",
            email: "email@example.com",
            password: "examplePassword",
            password_confirmation: "examplePassword"
            )

User#authenticate(password) returns the user instance if the input password is correct, and returns false if the input password is incorrect.