Specify restricting access without a valid token
When a user sends a GET request to the admin dashboard path without a valid token in the request, the server should respond with a 404. When a user sends a GET request to the admin dashbord path with a valid token in the request, the server should respond with a 200.