Refactor verifying a token's signature
This commit simplifies the admin dashboard controller and student dashboard controller by bringing the logic for verifying a token's signature higher in the inheritance hierarchy in
to ApplicationController, since that functionality is shared among all controllers.

Also, since authorizing a user involves a different role for each dashboard, this commit changes the name of each dashboard's authorize filter to show that it is explicitly filtering for that role. So, the admin dashboard controller authorizes with a method called authorize_admin and the student dashboard controller authorizes with a method called authorize_student.