Enable users to stay logged in indefinitely
This commit makes the JSON web token never expire. This will enable the common "Remember me" functionality seen in major websites. It is a tradeoff in that it does enable a more seamless and web-standard user experience but makes it possible for an attacker to indefinitely use a stolen token. This risk seems low for a music teaching app and the tradeoff seems acceptable.

The expiration functionality hadn't yet been implemented in the app anyway, so there isn't any change to the functionality of the app and the only change is that there is now no expiration timestamp in the token.