Specify that the token should be in an HTTP header
Previously, the JSON web token was included as a query parameter, and now that RESTful endpoints for creating users are being made, having query parameters relating to both user attributes and a JSON web token would make the nesting of query parameters a bit awkward.

This commit changes the specs to specify that the JSON web token should be in a header in the HTTP request. This means that the query parameters will be focused on REST conventions (for example, when creating a new resource over HTTP such as a user, the query parameters will contain the user attributes nested under a user namespace and won't contain a token that is unrelated to the user).