Implement receiving the token in a header
This commit implements receiving the JSON web token in an HTTP header instead of in the query parameters. (See the previous commit message for more details on the reasoning behind this change).

The new functionality for creating a user resource via a POST to a web service endpoint has still not been implemented yet, but this commit makes the specs green for the existing endpoints for user dashboards.