Specify and implement Lesson and Asset models
A Lesson has a title attribute as a string and validates for the presence of that attribute. A Lesson has a text attribute as a string and does not validate for the presence of that attribute. A Lesson has many Assets, and an Asset belongs to a Lesson. An Asset has a storageURL attribute as a string and validates for the presence of that attribute.