Adjust create lesson endpoint to respond with id
Instead of responding with the lesson object and its attributes and associated assets when a lesson is created, only respond with the newly created lesson's id. The client can then use this id to make another RESTful request to '/lessons/:id' and get all of the lesson attributes and associations via that endpoint.