Update teacher creation spec for email
This commit specifies a change in the teacher creation process where instead of an admin creating a password for a new teacher, the teacher will get an email that enables them to set their own password. The spec now specifies that an email will be sent in response to an authorized POST to "/api/v1/users".