Implement a NotYetCreatedUser model
When an admin starts the process of creating a new user, the new user will receive an email where they can set their password, but the User model won't be created in the database until their password is set.

This commit implements an intermediate model in that process so that when the admin enters a first name, last name, and email in the form, that information is saved in the database along with a unique identifier. This unique identifier can be used in a link in the password setting email that is sent to the new user, so that the server will be able to securely determine that the person who accesses that link is indeed the person who is authorized to own the user account and thus is authorized to set the password for that account.

A NotYetCreatedUser instance has the attributes "first_name", "last_name", "email", and "uuid" as strings, and the attribute "role" which is an enum that maps to an integer in the database and can be set and accessed with the names "teacher" and "student".