Implement sending a new user email
When the client sends an authorized POST to '/api/v1/send_new_user_email', including the attributes 'first_name', 'last_name', 'role', and 'email' nested under a 'user' key in the params, the server should send an email to the provided email address. This email contains a unique link to a page that the client will need to implement, where the user will be able to set their password and create a User record in the database. If the email is sent, the server responds with a 204 which indicates success but doesn't provide any other content for the client to render.