Implement including email address in password set link
In order to facilitate the client logging in immediately when the user sets their password, the user's email address is now available in the client app so that the email and password can be used immediately to log in. This commit implements including the soon-to-be-created user's email in the password set URL so that it is available to the client app.