Specify setting a new user's password
The client should be able to POST to '/api/v1/set_password', with a "user" key in the params under which would be nested the attributes "password", "password_confirmation", and "uuid". The value under "uuid" must be equal to the uuid attribute of a NotYetCreatedUser record. The server should respond to this request by creating a User record, destroying the NotYetCreatedUser record, and responding with a 204.