Specify a request to initiate a password reset
A client should be able to POST to '/api/v1/initiate_password_reset', with query parameters containing a user key under which is nested an email key and its value. If the provided email is associated with a user in the database, a PasswordReset instance should be created and an email should be sent that contains the PasswordReset instance's uuid.