Implement endpoint and email to initiate password reset
A client can POST to '/api/v1/initiate_password_reset', with query parameters containing a user key under which is nested an email key and its value. If the provided email is associated with a user in the database, a PasswordReset instance is created and an email is sent to the user. This email has a subject, and has a body containing the user's name as well as a query string containing the user's email and a uuid.