Specify a password reset endpoint
A client should be able to POST to '/api/v1/reset_password' with a user key in the query params whose value is a set of key-value pairs for a uuid, password, and password_confirmation. If the uuid can be used to find a PasswordReset record in the database, the associated user's password should be updated using the supplied password and password_confirmation.

If the user's password is updated, the server should delete the PasswordReset record and respond with a 204. If no user record is updated, the server should not delete a PasswordReset record and should respond with a 404.