Implement a password reset endpoint
A client can POST to '/api/v1/reset_password' with a user key in the query params whose value is a set of key-value pairs for a uuid, password, and password_confirmation. If the uuid can be used to find a PasswordReset record in the database, the associated user's password is updated using the supplied password and password_confirmation.

If the user's password is updated, the server deletes the PasswordReset record and responds with a 204. If no user record is updated, the server does not delete a PasswordReset record and responds with a 404.