Specify assignments to join lessons and students
This commit specifies an Assignment model that is a join model for a many-to-many relationship between lessons and users, with the intention that users will specifically be users who have the role 'student'. An assignment should have a lesson_id and a user_id. A lesson should be able to have many students through assignments (accessible with the instance method Lesson#students). A user should be able to have many lessons through assignments.