Implement assignments to join lessons and students
The Assignment model is a join model for a many-to-many relationship between lessons and users, with the intention that users will specifically be users who have the role 'student'. An assignment has a lesson_id and a user_id. A lesson can have many students through assignments (accessible with the instance method Lesson#students). A user can have many lessons through assignments.