Set up database seeding for development
To populate the database with seed data, run:
  $ rails db:seed

The database seeder creates several random lessons, assets, and users. It also creates an admin with the email 'admin@example.com', a teacher with the email 'teacher@example.com', and a student with the email 'student@example.com'. Each of these users has the easy-to-remember password 'asdf'.