Refactor generating JSON web tokens in specs
This commit introduces a helper module for the specs, SpecHelper, that defines a module method generate_token which takes a user object and returns a JSON web token. By calling this method to generate tokens in the specs, a lot of code that was not directly relevant to the feature being tested has now been removed. This makes the specs more focused and readable.