Implement an endpoint for assigned lessons
A client who has credentials as a Student user can send a GET to '/api/v1/assignments', with a "student_id" key-value pair in the query parameters, and receive a JSON response containing a "lesson_ids" key. The value under this key is a collection of the ids of the lessons assigned to that student.