Fix library loading in production
Passenger was showing an error in production where not all of the application source code had been loaded at runtime. Eager-loading the application source code solves this problem.

By default, Rails eager-loads everything in the app/ directory. This commit moves the lib/ directory and its contents into within the app/ directory so that all of the libraries are loaded when the app runs in production.